moving files around to follow suggested angular patterns
